Refactor assertSexp() to a test helper module
